Merge code for HGNC dictionary using R (v4.1.1)
Line 1: Working directory set for the session.
Line 3: Opening of the file containing the APP interactors. Parameters set for headers to be included and delimited for tab-separated values. Set as 'file.1'.
Line 5: Opening of the HGNC Gene Nomenclature dictionary. Parameters set for headers to be included and delimited for tab-separated values. Set as 'file.2'.
Line 7: Both tables merged to produce a 'results' table with HGNC dictionary information added to the corresponding APP interactors on the basis of the first column values being matched. 'All.x' parameter set to indicate the absence of a matched value with "n/a".
Line 9: 'Results' table exported as a tab-separated text file without quotes or line numbers into the working directory.