Create Semantic dictionary conversion using R (v4.1.1)
Line 1: Opening of the file containing the Ab42 g:Profiler enriched terms. Parameters set for headers to be included and delimited for tab-separated values. Set as 'file.1'.
Line 3: Opening of the semantic classification dictionary. Parameters set for headers to be included and delimited for tab-separated values. Set as 'file.2'.
Line 5: Both tables merged to produce a 'results' table with semantic dictionary classification added to the corresponding Ab42 enriched terms on the basis of the first column values being matched. 'All.x' parameter set to indicate the absence of a matched value with "n/a".
Line 7: 'Results' table exported as a tab-separated text file without quotes or line numbers into the working directory.